First draft for delete worker page
